requirements.txt
Flask
psycopg2-binary
gunicorn
python-dotenv
flask-session
requests
cryptography
